reworking things in Function Mgr so it does not call Document.ExpireSolution
